# Conflicts:
#	CHANGELOG.md
#	package-lock.json
#	package.json

+++ src/modules/billing/services/batch-service.js
@@ -1,17 +1,20 @@
 'use strict';
 
+const { partialRight } = require('lodash');
+
 const newRepos = require('../../../lib/connectors/repos');
 const mappers = require('../mappers');
 const { BATCH_STATUS } = require('../../../lib/models/batch');
 const { logger } = require('../../../logger');
 const Event = require('../../../lib/models/event');
 const eventService = require('../../../lib/services/events');
-const { BatchStatusError, BillingVolumeStatusError } = require('../lib/errors');
+const { BatchStatusError } = require('../lib/errors');
 const { NotFoundError } = require('../../../lib/errors');
 const { INCLUDE_IN_SUPPLEMENTARY_BILLING } = require('../../../lib/models/constants');
 const chargeModuleBillRunConnector = require('../../../lib/connectors/charge-module/bill-runs');
 
 const Batch = require('../../../lib/models/batch');
+const validators = require('../../../lib/models/validators');
 
 const invoiceLicenceService = require('./invoice-licences-service');
 const transactionsService = require('./transactions-service');
@@ -19,6 +22,8 @@ const billingVolumesService = require('./billing-volumes-service');
 const invoiceService = require('./invoice-service');
 const licencesService = require('../../../lib/services/licences');
 const config = require('../../../../config');
+
+const chargeModuleDecorators = require('../mappers/charge-module-decorators');
 
 /**
  * Loads a Batch instance by ID
@@ -61,16 +66,24 @@ const saveEvent = (type, status, user, batch) => {
 };
 
 const deleteBatch = async (batch, internalCallingUser) => {
+  if (batch.statusIsOneOf(Batch.BATCH_STATUS.sent)) {
+    throw new BatchStatusError(`Sent batch ${batch.id} cannot be deleted`);
+  }
+
   try {
     await licencesService.updateIncludeInSupplementaryBillingStatusForUnsentBatch(batch.id);
     await chargeModuleBillRunConnector.delete(batch.externalId);
 
+    // These are populated at every stage in the bill run
     await newRepos.billingBatchChargeVersionYears.deleteByBatchId(batch.id);
     await newRepos.billingBatchChargeVersions.deleteByBatchId(batch.id);
-    await newRepos.billingTransactions.deleteByBatchId(batch.id);
     await newRepos.billingVolumes.deleteByBatchId(batch.id);
+
+    // These tables are not yet populated at review stage in TPT
+    await newRepos.billingTransactions.deleteByBatchId(batch.id);
     await newRepos.billingInvoiceLicences.deleteByBatchId(batch.id);
-    await newRepos.billingInvoices.deleteByBatchId(batch.id);
+    await newRepos.billingInvoices.deleteByBatchId(batch.id, false);
+
     await newRepos.billingBatches.delete(batch.id);
 
     await saveEvent('billing-batch:cancel', 'delete', internalCallingUser, batch);
@@ -82,8 +95,24 @@ const deleteBatch = async (batch, internalCallingUser) => {
   }
 };
 
+/**
+ * Sets the batch status
+ * @param {String} batchId
+ * @param {BATCH_STATUS} status
+ * @param {BATCH_ERROR_CODE} [errorCode]
+ */
 const setStatus = (batchId, status) =>
   newRepos.billingBatches.update(batchId, { status });
+
+/**
+ * Sets the specified batch to 'error' status
+ *
+ * @param {String} batchId
+ * @param {BATCH_ERROR_CODE} errorCode The origin of the failure
+ * @return {Promise}
+ */
+const setErrorStatus = (batchId, errorCode) =>
+  newRepos.billingBatches.update(batchId, { status: Batch.BATCH_STATUS.error, errorCode });
 
 const approveBatch = async (batch, internalCallingUser) => {
   try {
@@ -100,20 +129,6 @@ const approveBatch = async (batch, internalCallingUser) => {
     await saveEvent('billing-batch:approve', 'error', internalCallingUser, batch);
     throw err;
   }
-};
-
-/**
- * Sets the specified batch to 'error' status
- *
- * @param {String} batchId
- * @param {BATCH_ERROR_CODE} errorCode The origin of the failure
- * @return {Promise}
- */
-const setErrorStatus = (batchId, errorCode) => {
-  return newRepos.billingBatches.update(batchId, {
-    status: Batch.BATCH_STATUS.error,
-    errorCode
-  });
 };
 
 const saveInvoiceLicenceTransactions = async (batch, invoice, invoiceLicence) => {
@@ -140,10 +155,15 @@ const saveInvoicesToDB = async batch => {
   }
 };
 
+/**
+ * Decorates the supplied batch with data retrieved from the charge module
+ * @param {Batch} batch
+ * @return {Promise<Batch>}
+ */
 const decorateBatchWithTotals = async batch => {
   try {
-    const { billRun: { summary } } = await chargeModuleBillRunConnector.get(batch.externalId);
-    batch.totals = mappers.totals.chargeModuleBillRunToBatchModel(summary);
+    const cmResponse = await chargeModuleBillRunConnector.get(batch.externalId);
+    return chargeModuleDecorators.decorateBatch(batch, cmResponse);
   } catch (err) {
     logger.info('Failed to decorate batch with totals. Waiting for CM API', err);
   }
@@ -151,17 +171,40 @@ const decorateBatchWithTotals = async batch => {
 };
 
 /**
- * Updates water.billing_batches with summary info from the charge module
+ * Persists the batch totals to water.billing_batches
  * @param {Batch} batch
  * @return {Promise}
  */
-const refreshTotals = async batch => {
-  const { billRun: { summary } } = await chargeModuleBillRunConnector.get(batch.externalId);
-  return newRepos.billingBatches.update(batch.id, {
-    invoiceCount: summary.invoiceCount,
-    creditNoteCount: summary.creditNoteCount,
-    netTotal: summary.netTotal
-  });
+const persistTotals = batch => {
+  const changes = batch.totals.pick([
+    'invoiceCount',
+    'creditNoteCount',
+    'netTotal'
+  ]);
+  return newRepos.billingBatches.update(batch.id, changes);
+};
+
+/**
+ * Updates water.billing_batches with summary info from the charge module
+ * and updates the is_deminimis flag for water.billing_transactions
+ * @param {Batch} batch
+ * @return {Promise}
+ */
+const refreshTotals = async batchId => {
+  validators.assertId(batchId);
+
+  // Load batch and map to service models
+  const data = await newRepos.billingBatches.findOneWithInvoicesWithTransactions(batchId);
+  const batch = mappers.batch.dbToModel(data);
+
+  // Load CM data and decorate service models
+  const cmResponse = await chargeModuleBillRunConnector.get(batch.externalId);
+  chargeModuleDecorators.decorateBatch(batch, cmResponse);
+
+  return Promise.all([
+    transactionsService.persistDeMinimis(batch),
+    persistTotals(batch)
+  ]);
 };
 
 /**
@@ -185,12 +228,8 @@ const getTransactionStatusCounts = async batchId => {
  * @returns {Batch} The updated batch if no transactions
  */
 const setStatusToEmptyWhenNoTransactions = async batch => {
-  const remainingTransactions = await newRepos.billingTransactions.findByBatchId(batch.id);
-
-  if (remainingTransactions.length === 0) {
-    return setStatus(batch.id, BATCH_STATUS.empty);
-  }
-  return batch;
+  const count = await newRepos.billingTransactions.countByBatchId(batch.id);
+  return count === 0 ? setStatus(batch.id, BATCH_STATUS.empty) : batch;
 };
 
 /**
@@ -254,19 +293,6 @@ const createChargeModuleBillRun = async batchId => {
   return batch.pickFrom(row, ['externalId', 'billRunNumber']);
 };
 
-const assertNoBillingVolumesWithTwoPartError = async batch => {
-  const billingVolumesWithTwoPartError = await billingVolumesService.getVolumesWithTwoPartError(batch);
-  if (billingVolumesWithTwoPartError.length > 0) {
-    throw new BillingVolumeStatusError('Cannot approve review. There are outstanding two part tariff errors to resolve');
-  }
-};
-
-const assertBatchStatusIsReview = batch => {
-  if (batch.status !== BATCH_STATUS.review) {
-    throw new BatchStatusError('Cannot approve review. Batch status must be "review"');
-  }
-};
-
 /**
  * Validates batch & transactions, then updates batch status to "processing"
  *
@@ -279,8 +305,10 @@ const assertBatchStatusIsReview = batch => {
  * @return {Promise<Batch>} resolves with Batch service model
  */
 const approveTptBatchReview = async batch => {
-  await assertNoBillingVolumesWithTwoPartError(batch);
-  assertBatchStatusIsReview(batch);
+  if (!batch.canApproveReview()) {
+    throw new BatchStatusError('Cannot approve review. Batch status must be "review"');
+  }
+  await billingVolumesService.approveVolumesForBatch(batch);
   await setStatus(batch.id, BATCH_STATUS.processing);
   return getBatchById(batch.id);
 };
@@ -357,6 +385,7 @@ exports.refreshTotals = refreshTotals;
 exports.saveInvoicesToDB = saveInvoicesToDB;
 exports.setErrorStatus = setErrorStatus;
 exports.setStatus = setStatus;
+exports.setStatusToReview = partialRight(setStatus, Batch.BATCH_STATUS.review);
 exports.setStatusToEmptyWhenNoTransactions = setStatusToEmptyWhenNoTransactions;
 exports.cleanup = cleanup;
 exports.create = create;
